Desktop/requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.29.0
autopep8==2.0.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
dnspython==2.3.0
Eel==0.14.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
pycodestyle==2.10.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pymongo==4.3.3
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyqt5-tools==5.15.4.3.2
python-dotenv==0.21.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
